[1136] Move vacancy code to a vacancy helper
